refactor: improve chat UX and SQL editor gutter
- Increase SQL editor gutter width with horizontal padding for better readability
- Move chat suggestions from empty state to separate bottom section
- Improve SQL execution flow: use openQuery for run action vs insertAtCursor for insert action
- Ensure suggestions are only shown when conversation is empty
