+++ apps/web/src/components/workspace/chat/chat-message-list.tsx
@@ -8,22 +8,14 @@ import {
   ConversationEmptyState,
   ConversationScrollButton,
 } from "@/components/ai-elements/conversation";
-import { Suggestion, Suggestions } from "@/components/ai-elements/suggestion";
 
 import { ChatMessage } from "./chat-message";
-
-const SQL_SUGGESTIONS = [
-  "Show all tables",
-  "Count rows in each table",
-  "Describe the schema",
-];
 
 interface ChatMessageListProps {
   messages: ChatMessageType[];
   connectionName: string;
   onInsertSql?: (sql: string) => void;
   onRunSql?: (sql: string) => void;
-  onSendSuggestion?: (text: string) => void;
 }
 
 export const ChatMessageList = ({
@@ -31,7 +23,6 @@ export const ChatMessageList = ({
   connectionName,
   onInsertSql,
   onRunSql,
-  onSendSuggestion,
 }: ChatMessageListProps) => (
   <Conversation className="flex-1">
     <ConversationContent>
@@ -40,15 +31,7 @@ export const ChatMessageList = ({
           icon={<MessageSquare className="size-6" />}
           title="Ask a question about your data"
           description={`Describe what you need in plain English and get SQL queries for ${connectionName}`}
-        >
-          {onSendSuggestion && (
-            <Suggestions className="mt-4 flex-wrap justify-center">
-              {SQL_SUGGESTIONS.map((s) => (
-                <Suggestion key={s} suggestion={s} onClick={onSendSuggestion} />
-              ))}
-            </Suggestions>
-          )}
-        </ConversationEmptyState>
+        />
       ) : (
         messages.map((msg) => (
           <ChatMessage
