+++ apps/web/src/components/workspace/sql-editor.tsx
@@ -151,7 +151,7 @@ export const SqlEditor = ({
         highlightActiveLine: true,
         lineNumbers: true,
       }}
-      className="h-full overflow-auto text-sm [&_.cm-editor]:h-full [&_.cm-scroller]:h-full [&_.cm-editor]:bg-background! [&_.cm-gutters]:bg-background! [&_.cm-gutters]:border-r-0! [&_.cm-activeLineGutter]:bg-background!"
+      className="h-full overflow-auto text-sm [&_.cm-editor]:h-full [&_.cm-scroller]:h-full [&_.cm-editor]:bg-background! [&_.cm-gutters]:bg-background! [&_.cm-gutters]:border-r-0! [&_.cm-gutters]:px-2! [&_.cm-activeLineGutter]:bg-background!"
     />
   );
 };
